feat: better install script

+++ install.sh
@@ -1,17 +1,18 @@
 #!/bin/bash
 
-# Install necessary deps
+echo "[*] Installing necessary dependencies ..."
 apt-get install curl zsh
-# Install oh-my-zsh
-curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh | sh; zsh
-# Sets zsh as default shell
+echo "[*] Installing and configuring zsh ..."
+curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh | sh
 sudo usermod --shell $(which zsh) $USER
 
-# Setups the plain theme
+echo "[*] Configuring zsh theme ..."
+[[ -f ~/.oh-my-zsh/themes/plain.zsh-theme ]] && rm ~/.oh-my-zsh/themes/plain.zsh-theme
 ln -s ~/.dotfiles/themes/plain.zsh-theme ~/.oh-my-zsh/themes/plain.zsh-theme
-# Intall the plugins
+echo "[*] Installing zsh plugins ..."
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 git clone https://github.com/zsh-users/zsh-autosuggestions $ZSH_CUSTOM/plugins/zsh-autosuggestions
-# Copies the config
-rm ~/.zshrc
+echo "[*] Configuring zsh ..."
+[[ -f ~/.zshrc ]] && rm ~/.zshrc
 ln -s ~/.dotfiles/.zshrc ~/.zshrc
+echo "[*] Install complete."
